requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
psycopg2-binary==2.9.9
spacy==3.7.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
pypdf2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.0
email-validator==2.1.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
